Clean up processor files
